HttpRequestHandler moved to Core namespace
